+++ packages/compiler-cli/BUILD.bazel
@@ -84,11 +84,11 @@ ts_library(
         "//packages/compiler-cli/src/ngtsc/program_driver",
         "//packages/compiler-cli/src/ngtsc/reflection",
         "//packages/compiler-cli/src/ngtsc/shims",
+        "//packages/compiler-cli/src/ngtsc/transform/jit",
         "//packages/compiler-cli/src/ngtsc/translator",
         "//packages/compiler-cli/src/ngtsc/typecheck",
         "//packages/compiler-cli/src/ngtsc/typecheck/api",
         "//packages/compiler-cli/src/ngtsc/util",
-        "//packages/compiler-cli/src/transformers/jit_transforms",
         "@npm//@bazel/concatjs",
         "@npm//@types/node",
         "@npm//@types/yargs",
