+++ packages/compiler-cli/src/ngtsc/core/BUILD.bazel
@@ -34,6 +34,7 @@ ts_library(
         "//packages/compiler-cli/src/ngtsc/shims",
         "//packages/compiler-cli/src/ngtsc/shims:api",
         "//packages/compiler-cli/src/ngtsc/transform",
+        "//packages/compiler-cli/src/ngtsc/transform/jit",
         "//packages/compiler-cli/src/ngtsc/typecheck",
         "//packages/compiler-cli/src/ngtsc/typecheck/api",
         "//packages/compiler-cli/src/ngtsc/typecheck/diagnostics",
@@ -44,7 +45,6 @@ ts_library(
         "//packages/compiler-cli/src/ngtsc/util",
         "//packages/compiler-cli/src/ngtsc/validation",
         "//packages/compiler-cli/src/ngtsc/xi18n",
-        "//packages/compiler-cli/src/transformers/jit_transforms",
         "@npm//@types/semver",
         "@npm//semver",
         "@npm//typescript",
