+++ packages/compiler-cli/src/ngtsc/transform/jit/src/initializer_api_transforms/input_function.ts
@@ -9,7 +9,7 @@
 import {core} from '@angular/compiler';
 import ts from 'typescript';
 
-import {isAngularDecorator, tryParseSignalInputMapping} from '../../../ngtsc/annotations';
+import {isAngularDecorator, tryParseSignalInputMapping} from '../../../../annotations';
 
 import {
   castAsAny,
