+++ packages/compiler-cli/src/ngtsc/transform/jit/src/initializer_api_transforms/query_functions.ts
@@ -11,7 +11,7 @@ import {
   queryDecoratorNames,
   QueryFunctionName,
   tryParseSignalQueryFromInitializer,
-} from '../../../ngtsc/annotations';
+} from '../../../../annotations';
 
 import {
   castAsAny,
