+++ packages/compiler-cli/src/ngtsc/transform/jit/test/BUILD.bazel
@@ -1,0 +1,27 @@
+load("//tools:defaults.bzl", "jasmine_node_test", "ts_library")
+
+ts_library(
+    name = "test_lib",
+    testonly = True,
+    srcs = [
+        "downlevel_decorators_transform_spec.ts",
+        "initializer_api_transforms_spec.ts",
+        "signal_queries_metadata_transform_spec.ts",
+    ],
+    deps = [
+        "//packages/compiler-cli/src/ngtsc/imports",
+        "//packages/compiler-cli/src/ngtsc/partial_evaluator",
+        "//packages/compiler-cli/src/ngtsc/reflection",
+        "//packages/compiler-cli/src/ngtsc/transform/jit",
+        "//packages/compiler-cli/test:test_utils",
+        "@npm//typescript",
+    ],
+)
+
+jasmine_node_test(
+    name = "test",
+    bootstrap = ["//tools/testing:node"],
+    deps = [
+        ":test_lib",
+    ],
+)
