+++ packages/core/test/acceptance/authoring/BUILD.bazel
@@ -24,7 +24,7 @@ ts_library(
         "//packages/compiler-cli/src/ngtsc/imports",
         "//packages/compiler-cli/src/ngtsc/partial_evaluator",
         "//packages/compiler-cli/src/ngtsc/reflection",
-        "//packages/compiler-cli/src/transformers/jit_transforms",
+        "//packages/compiler-cli/src/ngtsc/transform/jit",
         "@npm//typescript",
     ],
 )
